requirements.txt
torch>=1.12.1+cu113
Pillow>=9.3.0
pandas>=1.3.5
transformers>=4.25.1
torchmetrics>=0.11.0